fix: remove and update applications endpoints
BREAKING CHANGE: removed deprecated application endpoints
