+++ package.json
@@ -10,27 +10,27 @@
     }
   },
   "dependencies": {
-    "axios": "1.7.7",
-    "tslib": "2.6.2"
+    "axios": "1.7.8",
+    "tslib": "2.8.1"
   },
   "description": "A TypeScript SDK for the SentinelOne API",
   "devDependencies": {
     "@digitalroute/cz-conventional-changelog-for-jira": "8.0.1",
     "@pliancy/eslint-config-ts": "1.1.1",
     "@pliancy/semantic-release-config-npm": "2.2.0",
-    "@types/jest": "29.5.12",
-    "@types/node": "20.11.25",
-    "commitizen": "4.3.0",
+    "@types/jest": "29.5.14",
+    "@types/node": "22.10.1",
+    "commitizen": "4.3.1",
+    "concurrently": "9.1.0",
     "cpy-cli": "5.0.0",
-    "husky": "9.0.11",
+    "husky": "9.1.7",
     "jest": "29.7.0",
-    "jest-mock-axios": "4.7.3",
-    "npm-run-all2": "5.0.0",
+    "jest-mock-axios": "4.8.0",
     "open-cli": "7.2.0",
     "pinst": "3.0.0",
-    "rimraf": "5.0.5",
-    "ts-jest": "29.1.2",
-    "typescript": "5.4.2"
+    "rimraf": "6.0.1",
+    "ts-jest": "29.2.5",
+    "typescript": "5.7.2"
   },
   "engines": {
     "node": ">=18"
@@ -45,20 +45,20 @@
     "url": "https://github.com/pliancy/sentinelone-sdk"
   },
   "scripts": {
-    "build": "run-s -l clean test tsc copy",
+    "build": "yarn clean && yarn test && yarn tsc && yarn copy",
     "build:check": "tsc --noEmit",
     "clean": "rimraf coverage dist tmp",
     "copy": "pinst --disable && cpy package.json dist && cpy README.md dist && pinst --enable",
     "lint": "eslint \"src/**/*.ts\" --fix",
     "postinstall": "husky install",
-    "qa": "run-s -l lint test clean build:check",
+    "qa": "concurrently 'yarn lint' 'yarn test' 'yarn clean' 'yarn build:check'",
     "semantic-release": "semantic-release",
     "test": "jest",
     "tsc": "tsc -p tsconfig.build.json"
   },
   "version": "3.11.0",
   "volta": {
-    "node": "20.11.1",
+    "node": "22.11.0",
     "yarn": "1.22.22"
   }
 }
