+++ src/index.ts
@@ -7,4 +7,3 @@ export * from './lib/sentinel-one.types'
 export * from './lib/accounts/accounts.types'
 export * from './lib/settings/settings.types'
 export * from './lib/groups/groups.types'
-export * from './lib/applications/applications.types'
