+++ src/lib/sentinel-one.ts
@@ -8,7 +8,6 @@ import { Roles } from './roles/roles'
 import { Accounts } from './accounts/accounts'
 import { Settings } from './settings/settings'
 import { Groups } from './groups/groups'
-import { Applications } from './applications/applications'
 import { ServiceUsers } from './service-users/service-users'
 
 export class SentinelOne {
@@ -20,7 +19,6 @@ export class SentinelOne {
     accounts: Accounts
     settings: Settings
     groups: Groups
-    applications: Applications
 
     private readonly httpAgent: AxiosInstance
 
@@ -34,6 +32,5 @@ export class SentinelOne {
         this.accounts = new Accounts(this.httpAgent)
         this.settings = new Settings(this.httpAgent)
         this.groups = new Groups(this.httpAgent)
-        this.applications = new Applications(this.httpAgent)
     }
 }
